# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "aurevaRewards API",
    "version": "1.0.0",
    "description": "REST API for the Aureva Rewards blockchain loyalty platform. User endpoints require a Bearer JWT from `POST /auth/login`. Merchant endpoints require an `x-api-key` header.",
    "contact": {
      "name": "Aureva Rewards Team"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Local development (current v1)"
    },
    {
      "url": "https://api.aurevaRewards.io/api/v1",
      "description": "Production (current v1)"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Local development legacy alias for v1"
    },
    {
      "url": "https://api.aurevaRewards.io/api",
      "description": "Production legacy alias for v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from `POST /auth/login`."
      },
      "merchantApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Merchant API key returned on registration (shown once)."
      }
    },
...
  },
  "paths": {
...
    "/rewards/issue": {
      "post": {
        "tags": [
          "Rewards"
        ],
        "summary": "Enqueue idempotent reward issuance",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idempotencyKey",
                  "walletAddress",
                  "amount",
                  "campaignId"
                ],
                "properties": {
                  "idempotencyKey": {
                    "type": "string"
                  },
                  "walletAddress": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "campaignId": {
                    "type": "integer"
                  },
                  "userId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Duplicate — already processed"
          },
          "202": {
            "description": "Job queued"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/rewards/distribute": {
      "post": {
        "tags": [
          "Rewards"
        ],
        "summary": "Distribute AUR tokens to a customer wallet",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "walletAddress",
                  "amount"
                ],
                "properties": {
                  "walletAddress": {
                    "type": "string",
                    "example": "GBBD47..."
                  },
                  "amount": {
                    "type": "number",
                    "example": 50
                  },
                  "campaignId": {
                    "type": "integer",
                    "example": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens distributed",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "txHash": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No trustline or validation error"
          },
          "403": {
            "description": "Campaign not owned by merchant"
          },
          "404": {
            "description": "Campaign not found"
          },
          "502": {
            "description": "Chain error"
          }
        }
      }
    },
...
    "/trustline/verify": {
      "post": {
        "tags": [
          "Trustline"
        ],
        "summary": "Check whether a wallet has an active AUR trustline",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "walletAddress"
                ],
                "properties": {
                  "walletAddress": {
                    "type": "string",
                    "example": "GBBD47..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trustline status",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "properties": {
                        "exists": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid wallet address"
          }
        }
      }
    },
...
    "/wallet/verify": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Verify wallet connection",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "publicKey"
                ],
                "properties": {
                  "publicKey": {
                    "type": "string"
                  },
                  "walletType": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Wallet info"
          },
          "400": {
            "description": "Invalid wallet"
          }
        }
      }
    },
    "/wallet/balance": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "Get authenticated user live AUR balance",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Balance"
          },
          "400": {
            "description": "No wallet linked"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
...
  }
}